flask/requirement.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.9.3
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
flake8==4.0.1
Flask==2.0.2
frozenlist==1.3.0
future==0.18.2
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
line-bot-sdk==2.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==6.0.1
platformdirs==2.4.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
requests==2.27.1
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
yarl==1.7.2
